Make Ps4Socket connection more reliable
Closes #2

If the PS4 has gone idle or something, it won't respond to the
TCP. So, we can send WAKEUP and LAUNCH, wait a tick, and then
connect. It doesn't add too huge of a delay, so we'll just always
do it (for now, at least). This does mean `accountId` is pretty
much a required arg in the config dict.