Support not erroring from wake() if the device is already awake
0.6.6

The original behavior is mostly for the cli tool. Sending a wake
packet doesn't hurt if it's already awake, and allowing this flow
is convenient if you want to use Waker to connect a socket. The
terminology is a bit awkward, but it's nice and symmetrical to do
this, and Waker also has extra connect retry logic