Finally fix socket reconnects (I hope)
0.6.10

Packet used a static crypto resource to know
when to encrypt and decrypt packets. This was
fine for the cli tool, but now that I'm using
the socket more, it needs to be refactored.
For now, I can support a single Socket at a
time by simply clearing the static resouces