Refactor device search handling to better respect flags
This uses a breaking API change: the old findWhen is now findFirst,
which builds on a new findWhen that continues searching until
the callback returns True.

We could have kept searching until it returns False so that
the change would not be breaking, but if using findWhen instead
of findFirst, we assume that you will be more likely to want
to act on each device filtered, so that should be the default
behavior (IE: without an explicit return true).

We also could have kept `findWhen` for legacy and call the new
findWhen `findIf` instead, but my Clojure background makes me
find that icky---`if` should be for one thing, and `when` should
be for many.

So, breaking API change, with commensurate major version change