Preliminary On-Screen Keyboard support (#44)
* Add experimental support for the OSK

Refs #29

Some things are still not perfectly reliable.
Includes a refactor of the PacketManager into a class. I'd like
to gradually refactor things like this....

* Refactor osk bits into a separate class

Now if you want to do OSK stuff, you need to get an instance of
OnScreenKeyboard via Device.getKeyboard(), and that thing manages
the OSK state. It even emits a 'close' event when dismissed,
whether by the .close() or .submit() methods, or by someone else.

* Add CLI support for osk

Closes #29

* Update the README for OSK