Fail fast on unexpected (unrecoverable) login errors
This fixes a potential hang in such a case.

We weren't doing this before, but I think it was mostly just a result of
incomplete support for passcode auth. Now that we actually have handlers
for passcode auth, the exitProcess switch is not necessary in this case.
(in fact, we might want to just drop it in general...)